CiviCase - Prevent CaseType.definition xml from being escaped
The Api correctly reads/writes CaseType as an array, but feeding it a canonical xml string would cause the asinine HTMLInputCoder to escape all the angle brackets, resulting in a fatal runtime error whenever using CiviCase.
